Move eventDisplay into events subfolder
