+++ client/components/events/eventsDisplay.js
@@ -1,13 +1,13 @@
 import React from 'react';
-import Map from '../containers/GoogleMapContainer';
-import Search from './meetupSearchBar';
-import EventList from './MeetupEventsList';
 import Split from 'grommet/components/Split';
 import Sidebar from 'grommet/components/Sidebar';
 import Box from 'grommet/components/Box';
 import Header from 'grommet/components/Header';
 import Title from 'grommet/components/Title';
-import HeaderContainer from '../containers/headerContainer';
+import Map from '../../containers/GoogleMapContainer';
+import Search from '../meetupSearchBar';
+import EventList from '../MeetupEventsList';
+import HeaderContainer from '../../containers/headerContainer';
 
 const EventsDisplay = () =>
   (
